requirements.txt
Django==4.1.1
black==22.8.0
dj-database-url==1.0.0
flake8==5.0.4
flake8-django==1.1.5
isort==5.10.1
mypy==0.981
pytest==7.1.3
pytest-django==4.5.2
python-decouple==3.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
pytest-cov==4.0.0
drf-yasg==1.21.4
gunicorn
psycopg2
factory_boy==3.2.1
pytest-factoryboy==2.5.0